+++ pkg/service/v0/config.go
@@ -5,11 +5,12 @@ import (
 
 	"github.com/go-chi/render"
 	"github.com/owncloud/ocis-ocs/pkg/service/v0/data"
+	"github.com/owncloud/ocis-ocs/pkg/service/v0/response"
 )
 
 // GetConfig renders the ocs config endpoint
 func (o Ocs) GetConfig(w http.ResponseWriter, r *http.Request) {
-	render.Render(w, r, DataRender(&data.ConfigData{
+	render.Render(w, r, response.DataRender(&data.ConfigData{
 		Version: "1.7",  // TODO get from env
 		Website: "ocis", // TODO get from env
 		Host:    "",     // TODO get from FRONTEND config
