+++ pkg/service/v0/service.go
@@ -1,22 +1,17 @@
 package svc
 
 import (
-	"crypto/rand"
-	"encoding/hex"
 	"net/http"
 
-	"github.com/cs3org/reva/pkg/user"
 	"github.com/go-chi/chi"
 	"github.com/go-chi/chi/middleware"
 	"github.com/go-chi/render"
 
-	"github.com/micro/go-micro/v2/client/grpc"
-	merrors "github.com/micro/go-micro/v2/errors"
 	"github.com/owncloud/ocis-ocs/pkg/config"
 	ocsm "github.com/owncloud/ocis-ocs/pkg/middleware"
 	"github.com/owncloud/ocis-ocs/pkg/service/v0/data"
+	"github.com/owncloud/ocis-ocs/pkg/service/v0/response"
 	"github.com/owncloud/ocis-pkg/v2/log"
-	storepb "github.com/owncloud/ocis-store/pkg/proto/v0"
 )
 
 // Service defines the extension handlers.
@@ -47,7 +42,7 @@ func NewService(opts ...Option) Service {
 		))
 		r.Use(ocsm.OCSFormatCtx) // updates request Accept header according to format=(json|xml) query parameter
 		r.Route("/v{version:(1|2)}.php", func(r chi.Router) {
-			r.Use(svc.VersionCtx) // stores version in context
+			r.Use(response.VersionCtx) // stores version in context
 			r.Route("/apps/files_sharing/api/v1", func(r chi.Router) {})
 			r.Route("/apps/notifications/api/v1", func(r chi.Router) {})
 			r.Route("/cloud", func(r chi.Router) {
@@ -58,6 +53,7 @@ func NewService(opts ...Option) Service {
 				})
 				r.Route("/users", func(r chi.Router) {
 					r.Get("/", svc.ListUsers)
+					r.Get("/{userid}", svc.GetUser)
 				})
 			})
 			r.Route("/config", func(r chi.Router) {
@@ -83,95 +79,5 @@ func (o Ocs) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 // NotFound uses ErrRender to always return a proper OCS payload
 func (o Ocs) NotFound(w http.ResponseWriter, r *http.Request) {
-	render.Render(w, r, ErrRender(MetaUnknownError.StatusCode, "please check the syntax. API specifications are here: http://www.freedesktop.org/wiki/Specifications/open-collaboration-services"))
-}
-
-// GetUser returns the currently logged in user
-func (o Ocs) GetUser(w http.ResponseWriter, r *http.Request) {
-	u, ok := user.ContextGetUser(r.Context())
-	if !ok {
-		render.Render(w, r, ErrRender(MetaBadRequest.StatusCode, "missing user in context"))
-		return
-	}
-
-	render.Render(w, r, DataRender(&data.User{
-		ID:          u.Username, // TODO userid vs username! implications for clients if we return the userid here? -> implement graph ASAP?
-		DisplayName: u.DisplayName,
-		Email:       u.Mail,
-	}))
-}
-
-// GetSigningKey returns the signing key for the current user. It will create it on the fly if it does not exist
-// The signing key is part of the user settings and is used by the proxy to authenticate requests
-// Currently, the username is used as the OC-Credential
-func (o Ocs) GetSigningKey(w http.ResponseWriter, r *http.Request) {
-	u, ok := user.ContextGetUser(r.Context())
-	if !ok {
-		o.logger.Error().Msg("missing user in context")
-		render.Render(w, r, ErrRender(MetaBadRequest.StatusCode, "missing user in context"))
-		return
-	}
-	c := storepb.NewStoreService("com.owncloud.api.store", grpc.NewClient())
-	res, err := c.Read(r.Context(), &storepb.ReadRequest{
-		Options: &storepb.ReadOptions{
-			Database: "proxy",
-			Table:    "signing-keys",
-		},
-		Key: u.Username,
-	})
-	if err == nil && len(res.Records) > 0 {
-		render.Render(w, r, DataRender(&data.SigningKey{
-			User:       u.Username,
-			SigningKey: string(res.Records[0].Value),
-		}))
-		return
-	}
-	if err != nil {
-		e := merrors.Parse(err.Error())
-		if e.Code == http.StatusNotFound {
-			o.logger.Debug().Str("username", u.Username).Msg("signing key not found")
-			// not found is ok, so we can continue and generate the key on the fly
-		} else {
-			o.logger.Err(err).Msg("error reading from store")
-			render.Render(w, r, ErrRender(MetaServerError.StatusCode, "error reading from store"))
-			return
-		}
-	}
-
-	// try creating it
-	key := make([]byte, 64)
-	_, err = rand.Read(key[:])
-	if err != nil {
-		o.logger.Error().Err(err).Msg("could not generate signing key")
-		render.Render(w, r, ErrRender(MetaServerError.StatusCode, "could not generate signing key"))
-		return
-	}
-	signingKey := hex.EncodeToString(key)
-
-	_, err = c.Write(r.Context(), &storepb.WriteRequest{
-		Options: &storepb.WriteOptions{
-			Database: "proxy",
-			Table:    "signing-keys",
-		},
-		Record: &storepb.Record{
-			Key:   u.Username,
-			Value: []byte(signingKey),
-			// TODO Expiry?
-		},
-	})
-	if err != nil {
-		o.logger.Error().Err(err).Msg("error writing key")
-		render.Render(w, r, ErrRender(MetaServerError.StatusCode, "could not persist signing key"))
-		return
-	}
-
-	render.Render(w, r, DataRender(&data.SigningKey{
-		User:       u.Username,
-		SigningKey: signingKey,
-	}))
-}
-
-// ListUsers lists the users
-func (o Ocs) ListUsers(w http.ResponseWriter, r *http.Request) {
-	render.Render(w, r, ErrRender(MetaUnknownError.StatusCode, "please check the syntax. API specifications are here: http://www.freedesktop.org/wiki/Specifications/open-collaboration-services"))
+	render.Render(w, r, response.ErrRender(data.MetaUnknownError.StatusCode, "please check the syntax. API specifications are here: http://www.freedesktop.org/wiki/Specifications/open-collaboration-services"))
 }
